# Improve mysql loadData performance