--- a/liquibase-core/src/main/java/liquibase/change/core/LoadDataChange.java
+++ b/liquibase-core/src/main/java/liquibase/change/core/LoadDataChange.java
@@ -266,7 +266,10 @@ public class LoadDataChange extends AbstractTableChange implements ChangeWithCol
     public SqlStatement[] generateStatements(Database database) {
         boolean databaseSupportsBatchUpdates = false;
         try {
-            databaseSupportsBatchUpdates = database.supportsBatchUpdates();
+            if (!(database instanceof MySQLDatabase)) {
+                //mysql supports batch updates, but the performance vs. the big insert is worse
+                databaseSupportsBatchUpdates = database.supportsBatchUpdates();
+            }
         } catch (DatabaseException e) {
             throw new UnexpectedLiquibaseException(e);
         }