@MichaelCkr 2022-03-03
Found this issue after writing the text above
https://github.com/liquibase/liquibase/issues/2519